// src/client/generated-response/AuthTestResponse.ts
///////////////////////////////////
// !!! DO NOT EDIT THIS FILE !!! //
///////////////////////////////////

import { SlackAPIResponse } from "../response";
export type AuthTestResponse = SlackAPIResponse & {
  app_id?: string;
  app_name?: string;
  bot_id?: string;
  enterprise_id?: string;
  error?: string;
  is_enterprise_install?: boolean;
  needed?: string;
  ok: boolean;
  provided?: string;
  team?: string;
  team_id?: string;
  url?: string;
  user?: string;
  user_id?: string;
};